support renderer editor and categories overrides
